+++ app/myconsants/constants.go
@@ -1,0 +1,6 @@
+package myconsants
+
+const (
+	// AppName is the name of the application
+	AppName = "Automatic Hoyolab Check-in"
+)