+++ app/myconsole/operations.go
@@ -1,27 +1,36 @@
 package myconsole
 
-import "github.com/gonutz/w32/v2"
-
-var CurrentConsole w32.HWND
+import (
+	"github.com/WeeraW/auto-hoyolab-checkin/app/servicelogger"
+	"github.com/gonutz/w32/v2"
+)
 
 func HideConsole() {
-	console := w32.GetConsoleWindow()
-	if console == 0 {
+	if CurrentConsole == 0 {
+		servicelogger.Warning("No console attached.")
 		return // no console attached
 	}
-	_, consoleProcID := w32.GetWindowThreadProcessId(console)
+	_, consoleProcID := w32.GetWindowThreadProcessId(CurrentConsole)
 	if w32.GetCurrentProcessId() == consoleProcID {
-		w32.ShowWindowAsync(console, w32.SW_HIDE)
+		w32.ShowWindowAsync(CurrentConsole, w32.SW_HIDE)
 	}
 }
 
 func ShowConsole() {
-	console := w32.GetConsoleWindow()
-	if console == 0 {
+	if CurrentConsole == 0 {
+		servicelogger.Warning("No console attached.")
 		return // no console attached
 	}
-	_, consoleProcID := w32.GetWindowThreadProcessId(console)
+	_, consoleProcID := w32.GetWindowThreadProcessId(CurrentConsole)
 	if w32.GetCurrentProcessId() == consoleProcID {
-		w32.ShowWindowAsync(console, w32.SW_SHOW)
+		w32.ShowWindowAsync(CurrentConsole, w32.SW_SHOW)
 	}
+}
+
+func AttachConsole() {
+	if CurrentConsole != 0 {
+		servicelogger.Warning("Console already attached.")
+		return // already attached
+	}
+	CurrentConsole = w32.GetConsoleWindow()
 }